Moved main function from postcode.js to cli.js Main.js primarily has to do with outputting things to the command line, so should be kept separate from the interface agnostic logic in postcode.js
