+++ cli.js
@@ -1,4 +1,9 @@
-import { main, log } from "./postcodes.js";
+import {
+  validatePostcode,
+  geocodePostcode,
+  getCoordinates,
+  log,
+} from "./postcodes.js";
 
 const base_maps_url = "http://maps.google.com/maps?z=12&t=m&q=loc:";
 
@@ -15,6 +20,17 @@ export function getPostcodeFromArgs() {
   throw new Error(
     "No postcode given. You can provide a UK postcode by running `node app.js <postcode>. Your postcode should not contain spaces.`"
   );
+}
+
+export async function main(postcodeFromArgs) {
+  const postcode = postcodeFromArgs;
+  const postcodeIsValid = await validatePostcode(postcode);
+  if (postcodeIsValid) {
+    const coordinates = await geocodePostcode(postcode);
+    return getCoordinates(coordinates);
+  } else {
+    throw new Error(`${postcode} is not a valid uk postcode.`);
+  }
 }
 
 const coordinates = await main(getPostcodeFromArgs());
